requirements.txt
# Core Dependencies
numpy>=1.24.0
opencv-python>=4.8.0
scipy>=1.11.0
pyyaml>=6.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0  # YOLOv8

# Video Processing
ffmpeg-python>=0.2.0
av>=10.0.0

# Tracking
filterpy>=1.4.5
scikit-image>=0.21.0
lap>=0.4.0  # Linear Assignment Problem solver

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
tqdm>=4.65.0

# Evaluation
pandas>=2.0.0
openpyxl>=3.1.0

# Testing
pytest>=7.3.0
pytest-cov>=4.1.0

# Code Quality
black>=23.3.0
flake8>=6.0.0
mypy>=1.3.0

# Documentation
sphinx>=6.2.0
sphinx-rtd-theme>=1.2.0
